Move classify emails to llm_utils.py
